Add prompt color and LS_COLORS for Emacs terminals and Xterm
The prompt color and LS_COLORS that we use for the Xterm (which has
DarkSlateGray background) do not look good if we use one of the
Emacs terminals.  Therefore, we apply a different set of colors
if $TERM == dumb (i.e. emacs shell) or $TERM == eterm-color (i.e.
emacs terminal) than we do for $TERM == xterm (X-terminal).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>